Setup dynamic slash command registration
